+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v1.1.0
+
+- Add a lots of configs to custom BigMouth behavior
+- Add a config 'canBeEveryItem', disabled by default. If enabled, BigMouth can transform into any scrap item, even modded ones !
+- You can also config which item BigMouth cannot transform into.
+
 ### v1.0.3
 
 - Set power level to 1 and max count to 4